Create card component file
This commit separates the card component into it's own file to improve code readability.
