+++ front-end/src/components/CardContainer.jsx
@@ -1,3 +1,4 @@
+import { Card } from "./Card"
 import "../styles/CardContainer.css"
 
 let testData = [
@@ -6,33 +7,8 @@ let testData = [
     {id: 3, title: "Title", type: "Celebration", author: "Johnny"},
 ]
 
-function ViewBoardButton(props) {
-    return (
-        <button onClick={props.onClick}>View Board</button>
-    )
-}
-function DeleteBoardButton(props) {
-    return (
-        <button onClick={props.onClick}>Delete Board</button>
-    )
-}
 
-function Card(props) {
-    return (
-        <div className="card">
-            <div className="card-information">
-                <img src="#" alt="" />
-                <h2 className="card-text">{props.title}</h2>
-                <h3 className="card-text">{props.type}</h3>
-            </div>
 
-            <div className="card-interactables">
-                <ViewBoardButton />
-                <DeleteBoardButton />
-            </div>
-        </div>
-    )
-}
 
 export function CardContainer(props) {
     return (
