+++ front-end/src/styles/CardContainer.css
@@ -32,51 +32,7 @@
     background-color: var(--bg-secondary);
 }
 
-.card {
-    display: flex;
-    flex-direction: column;
-    background-color: var(--bg-primary);
-    border: 1px solid var(--border-light);
-    box-shadow: 0 2px 8px var(--shadow-light);
-    transition: box-shadow 0.3s ease;
-    transition: transform 0.3s ease;
-    border-radius: 10px;
-}
 
-.card-information img {
-    height: 300px;
-    width: 200px;
-    border-radius: 10px;
-    margin-top: 7px;
-}
-
-.card-text {
-    margin: 0px;
-    margin-top: 5px;
-}
-
-.card-information {
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-}
-
-.card-interactables {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    border-radius: 10px;
-    margin-top: 10px;
-}
-
-.card-interactables button {
-    width: 45%;
-}
-
-
-.card:hover {
-    transform: translate(0, -5px);
-}
 
 
 @media (min-width: 480px) {
